+++ src/model/index.ts
@@ -20,11 +20,11 @@
 import { HeaderTemplate, InlineTemplate, JournalPageType, ScopedTemplate } from './config';
 import { InlineString } from './inline';
 import { Input, SelectedInput, NoteInput } from './input';
-import { TemplateInfo, InlineTemplate22 } from './templates';
+import { TemplateInfo, BaseDirectory } from './templates';
+import { FileEntry } from "./files";
 
 
 export {
-    Input, SelectedInput, NoteInput, TemplateInfo, InlineTemplate22, InlineString, InlineTemplate,
-    ScopedTemplate, JournalPageType,  HeaderTemplate
+    Input, SelectedInput, NoteInput, TemplateInfo, BaseDirectory, InlineString, InlineTemplate,
+    ScopedTemplate, JournalPageType,  HeaderTemplate, FileEntry
 };
-// export { PickDayItem } from './pickDayItem';
