+++ src/provider/features/show-pick-list.ts
@@ -1,0 +1,9 @@
+import * as vscode from 'vscode';
+import * as J from '../..';
+
+/**
+ * Feature responsible for scanning the journal and notes folders and collecting the items displayed in the picklist
+*/
+ 
+export class ShowPickLists {
+}
