+++ src/provider/index.ts
@@ -1,6 +1,7 @@
 export { MatchInput  } from "./features/match-input";
 export { SyncNoteLinks } from "./features/sync-note-links";
 export { LoadNotes  } from "./features/load-note";
+export { ScanEntries } from "./features/scan-entries";
 
 export * as Commands from './commands';
 
@@ -9,3 +10,4 @@ export { ShiftTaskCodeLens } from './codelens/shift-task';
 
 export { CompletedTaskActions } from './codeactions/for-completed-tasks';
 export { OpenTaskActions } from './codeactions/for-open-tasks';
+
