+++ src/util/controller.ts
@@ -27,7 +27,7 @@ export class Ctrl {
 
 
     private _config: J.Extension.Configuration;
-    private _ui: J.Extension.VSCode;
+    private _ui: J.Extension.Dialogues;
     private _parser: J.Actions.Parser;
     private _writer: J.Actions.Writer;
     private _reader: J.Actions.Reader;
@@ -45,7 +45,7 @@ export class Ctrl {
         this._writer = new J.Actions.Writer(this);
         this._reader = new J.Actions.Reader(this);
         this._inject = new J.Actions.Inject(this);
-        this._ui = new J.Extension.VSCode(this);
+        this._ui = new J.Extension.Dialogues(this);
     }
 
 
@@ -54,7 +54,7 @@ export class Ctrl {
      * Getter $ui
      * @return {J.Extension.VSCode}
      */
-    public get ui(): J.Extension.VSCode {
+    public get ui(): J.Extension.Dialogues {
         return this._ui;
     }
